Adding more test coverage for GeoCore input/output. Also fixed path conversion
Path conversion needed adjusting to match new Clipper behavior.
Added more tests for GeoCore I/O validation for entities and across formats, also extending/fixing the internal calls to make validation possible.

+++ Common/geoWrangler/inflate_p64.cs
@@ -81,12 +81,12 @@ public static partial class GeoWrangler
         return edge;
     }
     
-    public static Path64 inflatePath(Path64 source, int width)
+    public static Path64 inflatePath(Path64 source, double width)
     {
         return pInflatePath(source, width);
     }
 
-    private static Path64 pInflatePath(Path64 source, int width)
+    private static Path64 pInflatePath(Path64 source, double width)
     {
         switch (width)
         {
